// src/app/shared/models/error-message-model.ts
import { ErrorMessageData } from 'desy-angular';

export class ErrorMessageModel implements ErrorMessageData{
  ariaControls: string;
  ariaCurrent: string;
  ariaDescribedBy: string;
  ariaDisabled: string;
  ariaErrorMessage: string;
  ariaExpanded: string;
  ariaHasPopup: string;
  ariaHidden: string;
  ariaLabel: string;
  ariaLabelledBy: string;
  ariaLive: string;
  classes: string;
  html: string;
  id: string;
  role: string;
  text: string;
  visuallyHiddenText: string;
  tabindex: string;
}